상품을 생성하는 메소드 테스트

+++ app/src/test/java/com/codesoom/assignment/domain/ProductRepositoryTest.java
@@ -69,7 +69,7 @@ class ProductRepositoryTest {
                 List<Product> products = productRepository.findAll();
 
                 assertThat(products).isNotEmpty();
-                assertThat(products).hasSize((int) createProductSize);
+                assertThat(products).hasSize(createProductSize);
             }
         }
